WIP: SRP principles over layers
