+++ ConwaysGameOfLife.sln
@@ -12,6 +12,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		Directory.Build.props = Directory.Build.props
 		Directory.Packages.props = Directory.Packages.props
 		global.json = global.json
+		docker-compose.yml = docker-compose.yml
+		Makefile = Makefile
 	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{229E1888-0DAC-4F7C-A9D6-9AF3F94C293D}"
@@ -26,9 +28,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Conways.GameOfLife.Integrat
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Conways.GameOfLife.API", "src\Conways.GameOfLife.API\Conways.GameOfLife.API.csproj", "{C28745E5-796A-41A1-8BCF-9F11212E4C17}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Conways.GameOfLife.Infrastructure.PostgreSQL", "src\Conways.GameOfLife.Infrastructure.PostgreSQL\Conways.GameOfLife.Infrastructure.PostgreSQL.csproj", "{CE05B38E-EA36-445D-97B8-BC24B87C953E}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Conways.GameOfLife.Infrastructure", "src\Conways.GameOfLife.Infrastructure\Conways.GameOfLife.Infrastructure.csproj", "{B3ABB3FE-B01C-4B7E-B12B-62B8505FB9D2}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Conways.GameOfLife.Infrastructure", "src\Conways.GameOfLife.Infrastructure\Conways.GameOfLife.Infrastructure.csproj", "{CE05B38E-EA36-445D-97B8-BC24B87C953E}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".github", ".github", "{4D6B81F7-ACAF-484F-8D65-C291E5867C85}"
 	ProjectSection(SolutionItems) = preProject
@@ -80,7 +80,6 @@ Global
 		{F3E9AB37-6274-4933-9334-356E9030B94F} = {229E1888-0DAC-4F7C-A9D6-9AF3F94C293D}
 		{C28745E5-796A-41A1-8BCF-9F11212E4C17} = {F3E9AB37-6274-4933-9334-356E9030B94F}
 		{C419FB97-4442-4D28-A62D-16F25FFC1220} = {F3E9AB37-6274-4933-9334-356E9030B94F}
-		{B3ABB3FE-B01C-4B7E-B12B-62B8505FB9D2} = {F3E9AB37-6274-4933-9334-356E9030B94F}
 		{CE05B38E-EA36-445D-97B8-BC24B87C953E} = {F3E9AB37-6274-4933-9334-356E9030B94F}
 		{7DED051D-B336-4C32-9ECD-A034321A9D40} = {1F02AA9F-62AD-40D9-BD42-451C4E94CE31}
 		{C7F06B77-F1CC-47F4-B67D-10840416F9FC} = {1F02AA9F-62AD-40D9-BD42-451C4E94CE31}
@@ -107,10 +106,6 @@ Global
 		{CE05B38E-EA36-445D-97B8-BC24B87C953E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CE05B38E-EA36-445D-97B8-BC24B87C953E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CE05B38E-EA36-445D-97B8-BC24B87C953E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B3ABB3FE-B01C-4B7E-B12B-62B8505FB9D2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B3ABB3FE-B01C-4B7E-B12B-62B8505FB9D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B3ABB3FE-B01C-4B7E-B12B-62B8505FB9D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B3ABB3FE-B01C-4B7E-B12B-62B8505FB9D2}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7DED051D-B336-4C32-9ECD-A034321A9D40}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7DED051D-B336-4C32-9ECD-A034321A9D40}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7DED051D-B336-4C32-9ECD-A034321A9D40}.Release|Any CPU.ActiveCfg = Release|Any CPU
