+++ Directory.Packages.props
@@ -3,49 +3,50 @@
     <ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
   </PropertyGroup>
   <ItemGroup>
-    <PackageVersion Include="Asp.Versioning.Abstractions" Version="8.1.0"/>
-    <PackageVersion Include="Aspire.Hosting.AppHost" Version="9.2.1"/>
-    <PackageVersion Include="Aspire.Hosting.Testing" Version="9.2.1"/>
-    <PackageVersion Include="Asp.Versioning.Http" Version="8.1.0"/>
-    <PackageVersion Include="Asp.Versioning.Mvc.ApiExplorer" Version="8.1.0"/>
-    <PackageVersion Include="Aspire.Hosting.PostgreSQL" Version="9.2.1"/>
-    <PackageVersion Include="Aspire.Hosting.Redis" Version="9.2.1"/>
-    <PackageVersion Include="AspNetCore.HealthChecks.NpgSql" Version="9.0.0"/>
-    <PackageVersion Include="coverlet.collector" Version="6.0.2"/>
-    <PackageVersion Include="FluentAssertions" Version="[7.2.0]"/>
-    <PackageVersion Include="FluentValidation.AspNetCore" Version="11.3.0"/>
-    <PackageVersion Include="Hashids.net" Version="1.7.0"/>
-    <PackageVersion Include="MediatR" Version="12.5.0"/>
-    <PackageVersion Include="Microsoft.AspNetCore.Mvc.Testing" Version="9.0.5"/>
-    <PackageVersion Include="Microsoft.AspNetCore.OpenApi" Version="9.0.5"/>
-    <PackageVersion Include="Microsoft.EntityFrameworkCore.Analyzers" Version="9.0.5"/>
+    <PackageVersion Include="Asp.Versioning.Abstractions" Version="8.1.0" />
+    <PackageVersion Include="Aspire.Hosting.AppHost" Version="9.2.1" />
+    <PackageVersion Include="Aspire.Hosting.Testing" Version="9.2.1" />
+    <PackageVersion Include="Asp.Versioning.Http" Version="8.1.0" />
+    <PackageVersion Include="Asp.Versioning.Mvc.ApiExplorer" Version="8.1.0" />
+    <PackageVersion Include="Aspire.Hosting.PostgreSQL" Version="9.2.1" />
+    <PackageVersion Include="Aspire.Hosting.Redis" Version="9.2.1" />
+    <PackageVersion Include="AspNetCore.HealthChecks.NpgSql" Version="9.0.0" />
+    <PackageVersion Include="coverlet.collector" Version="6.0.2" />
+    <PackageVersion Include="FluentAssertions" Version="[7.2.0]" />
+    <PackageVersion Include="FluentValidation.AspNetCore" Version="11.3.0" />
+    <PackageVersion Include="Hashids.net" Version="1.7.0" />
+    <PackageVersion Include="MediatR" Version="12.5.0" />
+    <PackageVersion Include="Microsoft.AspNetCore.Mvc.Testing" Version="9.0.5" />
+    <PackageVersion Include="Microsoft.AspNetCore.OpenApi" Version="9.0.5" />
+    <PackageVersion Include="Microsoft.EntityFrameworkCore.Analyzers" Version="9.0.5" />
     <PackageVersion Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.5">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageVersion>
-    <PackageVersion Include="Microsoft.EntityFrameworkCore.Relational" Version="9.0.5"/>
-    <PackageVersion Include="Microsoft.Extensions.Configuration" Version="9.0.5"/>
-    <PackageVersion Include="Microsoft.Extensions.Configuration.Binder" Version="9.0.5"/>
-    <PackageVersion Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="9.0.5"/>
-    <PackageVersion Include="Microsoft.Extensions.Configuration.Json" Version="9.0.5"/>
-    <PackageVersion Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.5"/>
-    <PackageVersion Include="Microsoft.Extensions.ServiceDiscovery" Version="9.2.1"/>
-    <PackageVersion Include="Microsoft.Extensions.Http.Resilience" Version="9.5.0"/>
-    <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.13.0"/>
-    <PackageVersion Include="Microsoft.Testing.Extensions.CodeCoverage" Version="17.14.2"/>
-    <PackageVersion Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.4"/>
-    <PackageVersion Include="OpenTelemetry" Version="1.12.0"/>
-    <PackageVersion Include="OpenTelemetry.Exporter.Console" Version="1.12.0"/>
-    <PackageVersion Include="OpenTelemetry.Exporter.OpenTelemetryProtocol" Version="1.12.0"/>
-    <PackageVersion Include="OpenTelemetry.Extensions.Hosting" Version="1.12.0"/>
-    <PackageVersion Include="OpenTelemetry.Instrumentation.AspNetCore" Version="1.12.0"/>
-    <PackageVersion Include="OpenTelemetry.Instrumentation.EntityFrameworkCore" Version="1.12.0-beta.1"/>
-    <PackageVersion Include="OpenTelemetry.Instrumentation.EventCounters" Version="1.5.1-alpha.1"/>
-    <PackageVersion Include="OpenTelemetry.Instrumentation.Http" Version="1.12.0"/>
-    <PackageVersion Include="OpenTelemetry.Instrumentation.Runtime" Version="1.12.0"/>
-    <PackageVersion Include="Scalar.AspNetCore" Version="2.3.0"/>
-    <PackageVersion Include="Testcontainers.PostgreSql" Version="4.4.0"/>
-    <PackageVersion Include="xunit.v3" Version="2.0.2"/>
+    <PackageVersion Include="Microsoft.EntityFrameworkCore.Relational" Version="9.0.5" />
+    <PackageVersion Include="Microsoft.Extensions.Configuration" Version="9.0.5" />
+    <PackageVersion Include="Microsoft.Extensions.Configuration.Binder" Version="9.0.5" />
+    <PackageVersion Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="9.0.5" />
+    <PackageVersion Include="Microsoft.Extensions.Configuration.Json" Version="9.0.5" />
+    <PackageVersion Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.5" />
+    <PackageVersion Include="Microsoft.Extensions.ServiceDiscovery" Version="9.2.1" />
+    <PackageVersion Include="Microsoft.Extensions.Http.Resilience" Version="9.5.0" />
+    <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.13.0" />
+    <PackageVersion Include="Microsoft.Testing.Extensions.CodeCoverage" Version="17.14.2" />
+    <PackageVersion Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.4" />
+    <PackageVersion Include="OpenTelemetry" Version="1.12.0" />
+    <PackageVersion Include="OpenTelemetry.Exporter.Console" Version="1.12.0" />
+    <PackageVersion Include="OpenTelemetry.Exporter.OpenTelemetryProtocol" Version="1.12.0" />
+    <PackageVersion Include="OpenTelemetry.Extensions.Hosting" Version="1.12.0" />
+    <PackageVersion Include="OpenTelemetry.Instrumentation.AspNetCore" Version="1.12.0" />
+    <PackageVersion Include="OpenTelemetry.Instrumentation.EntityFrameworkCore" Version="1.12.0-beta.1" />
+    <PackageVersion Include="OpenTelemetry.Instrumentation.EventCounters" Version="1.5.1-alpha.1" />
+    <PackageVersion Include="OpenTelemetry.Instrumentation.Http" Version="1.12.0" />
+    <PackageVersion Include="OpenTelemetry.Instrumentation.Runtime" Version="1.12.0" />
+    <PackageVersion Include="Scalar.AspNetCore" Version="2.3.0" />
+    <PackageVersion Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
+    <PackageVersion Include="Testcontainers.PostgreSql" Version="4.4.0" />
+    <PackageVersion Include="xunit.v3" Version="2.0.2" />
     <PackageVersion Include="xunit.runner.visualstudio" Version="3.1.0">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
