+++ src/Conways.GameOfLife.API/Extensions/ServiceCollectionExtensions.cs
@@ -1,27 +1,25 @@
-using System.Diagnostics.CodeAnalysis;
 using Conways.GameOfLife.API.Behaviors;
-using Conways.GameOfLife.Infrastructure.PostgreSQL;
+using Conways.GameOfLife.Infrastructure;
+using Conways.GameOfLife.Infrastructure.Persistence;
 using FluentValidation;
 
 namespace Conways.GameOfLife.API.Extensions;
 
-[ExcludeFromCodeCoverage]
 public static class ServiceCollectionExtensions
 {
-    public static IServiceCollection AddCQRS(this IServiceCollection services)
+    internal static IServiceCollection AddCQRS(this IServiceCollection services)
     {
-        return services.AddMediatR(config =>
+        services.AddMediatR(config =>
         {
             config.RegisterServicesFromAssemblyContaining<Program>();
             config.AddOpenBehavior(typeof(LoggingBehavior<,>));
             config.AddOpenBehavior(typeof(ValidationBehavior<,>));
             config.Lifetime = ServiceLifetime.Scoped;
         });
-    }
 
-    public static IServiceCollection AddFluentValidators(this IServiceCollection services)
-    {
-        return services.AddValidatorsFromAssemblyContaining<Program>();
+        services.AddValidatorsFromAssemblyContaining<Program>();
+
+        return services;
     }
 
     internal static IServiceCollection AddPostgreHealthCheck(this IServiceCollection services,
