+++ src/Conways.GameOfLife.API/Features/CreateBoard/CreateBoardCommandHandler.cs
@@ -1,6 +1,7 @@
 using Conways.GameOfLife.Domain;
 using Conways.GameOfLife.Infrastructure.Extensions;
-using Conways.GameOfLife.Infrastructure.PostgreSQL;
+using Conways.GameOfLife.Infrastructure;
+using Conways.GameOfLife.Infrastructure.Persistence;
 using HashidsNet;
 using MediatR;
