--- src/Conways.GameOfLife.Infrastructure.PostgreSQL/BoardDesignTimeDbContextFactory.cs
@@ -1,29 +1,0 @@
-using System.Diagnostics.CodeAnalysis;
-using Conways.GameOfLife.Infrastructure.Factories;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Design;
-using Microsoft.Extensions.Configuration;
-
-namespace Conways.GameOfLife.Infrastructure.PostgreSQL;
-
-[ExcludeFromCodeCoverage]
-public class BoardDesignTimeDbContextFactory : IDesignTimeDbContextFactory<BoardDbContext>
-{
-    public BoardDbContext CreateDbContext(string[] args)
-    {
-        var configuration = ConfigurationFactory.CreateConfiguration();
-
-        var connectionString = configuration.GetConnectionString(BoardDbContext.DatabaseName);
-
-        if (string.IsNullOrWhiteSpace(connectionString))
-        {
-            throw new InvalidOperationException($"Connection string with key 'Postgres' was not found");
-        }
-
-        var options = new DbContextOptionsBuilder<BoardDbContext>()
-            .UseNpgsql(connectionString, builder => builder.EnableRetryOnFailure(3))
-            .Options;
-
-        return new BoardDbContext(options);
-    }
-}
