--- src/Conways.GameOfLife.Infrastructure.PostgreSQL/Conways.GameOfLife.Infrastructure.PostgreSQL.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Analyzers"/>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational"/>
-    <PackageReference Include="Microsoft.Extensions.Configuration"/>
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL"/>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Conways.GameOfLife.Domain\Conways.GameOfLife.Domain.csproj"/>
-    <ProjectReference Include="..\Conways.GameOfLife.Infrastructure\Conways.GameOfLife.Infrastructure.csproj"/>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Migrations\"/>
-  </ItemGroup>
-
-</Project>
