+++ src/Conways.GameOfLife.Infrastructure/Persistence/BoardDbContext.cs
@@ -1,6 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 
-namespace Conways.GameOfLife.Infrastructure.PostgreSQL;
+namespace Conways.GameOfLife.Infrastructure.Persistence;
 
 public sealed class BoardDbContext : DbContext
 {
