+++ tests/Conways.GameOfLife.IntegrationTests/Features/FinalGeneration/FinalGenerationTests.cs
@@ -1,3 +1,5 @@
+using Conways.GameOfLife.Infrastructure.Persistence;
+
 namespace Conways.GameOfLife.IntegrationTests.Features.FinalGeneration;
 
 public class FinalGenerationTests
