+++ tests/Conways.GameOfLife.IntegrationTests/GlobalUsings.cs
@@ -5,8 +5,7 @@ global using Conways.GameOfLife.API.Features.NextGeneration;
 global using Conways.GameOfLife.API.Features.NextGenerations;
 global using Conways.GameOfLife.Domain.Exceptions;
 global using Conways.GameOfLife.Domain;
-global using Conways.GameOfLife.Infrastructure.PostgreSQL.Interceptors;
-global using Conways.GameOfLife.Infrastructure.PostgreSQL;
+global using Conways.GameOfLife.Infrastructure;
 global using Conways.GameOfLife.IntegrationTests.Factories;
 global using FluentAssertions;
 global using FluentValidation;
